A/S for Administration
Reviewed Sara Jackson’s corrections to the A/S for Administration list.

Joe — A/S for Administration is actually a good place to start, because
it includes several people whose titles changed mid-stream.

A few additional points: Sara designated Carlisle Humelsine and Loy
Henderson as deputy-under-secretary-of-state-for-administration-2011 —
I wonder if it doesn’t make more sense to treat this particular Deputy
U/S as a wholly separate entity, given the long stretch of emptiness
between the two.

Additionally, note that Stephen J. Rodriguez served as Acting Assistant
Secretary. Whereas we have a ready-made solution for charge d’affaires
ad interim, I don’t believe that we have one for long-term acting
principal officers. Will we need to create an “acting” position every
time this happens
(acting-assistant-secretary-of-state-for-administration,
acting-assistant-secretary-of-state-for-public-affairs, etc etc etc)?

Anyway, I’ll be interested to see what you think once you’ve reviewed
the material.