Arms Control
Corrected the XML for two positions: A/S for Arms Control,
Verification, and Compliance, and A/S for International Security and
Nonproliferation.

AVC is a straightforward case: starts as Verification and Compliance,
then becomes Verification, Compliance, and Implementation, and arrives
at Arms Control, Verification, and Comlpiance.

ISN was formed by the merger of two, now discontinued, bureaus: Bureau
of Arms Control and the Bureau of Non-Proliferation.